Starting to build out testing (#2)
* Fixing setup.cfg issue

* Literally just the beginnings of testing

* Adding configuration for github workflows

* Some issues with package installation -- working locally

* I think this works?

* Forgot to change version reference

* flake8 fix

* Adding pandas -- why did I remove it again?

* Filtering didn't work for code cov

* Maybe this will work

* Renaming the package to bayte

* Python 3.9

* Adding some more testing

* Additional testing

* Finishing up the tests for the encoder

* Adding a utility module for generating data with categoricals

* Changing the regression data generator utility and adding some tests

* flake8

* Small adjustment

* Additional utility testing

* Adding testing for the ensemble

* Refining the regression data testing

* Revamping the make regression function.

* Formatting
