+++ .coveragerc
@@ -1,0 +1,2 @@
+[report]
+show_missing = True
