+++ .github/workflows/test-package.yml
@@ -1,0 +1,45 @@
+# Install and test
+
+name: Python package
+
+on:
+    push:
+        branches:
+            - develop
+            - main
+    pull_request:
+        branches:
+            - develop
+            - main
+
+jobs:
+    test:
+        runs-on: ubuntu-latest
+        strategy:
+            matrix:
+                python_version:
+                    - 3.7
+                    - 3.8
+                    - 3.9
+        steps:
+            - uses: actions/checkout@v2
+            - name: Set up Python ${{ matrix.python-version }}
+              uses: actions/setup-python@v2
+              with:
+                python-version: ${{ matrix.python-version }}
+            - name: Install dependencies
+              run: python -m pip install -r requirements.txt .[dev]
+            - name: Check docstrings
+              run: python -m pydocstyle bayte --convention=numpy
+            - name: Check static typing
+              run: python -m mypy bayte
+            - name: Run flake8
+              run: python -m flake8 bayte
+            - name: Run unit testing
+              run: python -m pytest tests --cov=./ --cov-report=xml
+            - name: Run Codecov
+              uses: codecov/codecov-action@v2
+              with:
+                fail_ci_if_error: true
+                files: ./coverage.xml
+                verbose: true
