+++ bayte/_meta.py
@@ -1,0 +1,3 @@
+"""Package metadata."""
+
+__version__ = "0.1.0"
