+++ requirements.txt
@@ -8,11 +8,20 @@ joblib==1.1.0
     # via scikit-learn
 numpy==1.21.4
     # via
+    #   pandas
     #   scikit-learn
     #   scipy
+pandas==1.3.4
+    # via bayesian-target (setup.cfg)
+python-dateutil==2.8.2
+    # via pandas
+pytz==2021.3
+    # via pandas
 scikit-learn==1.0.1
     # via bayesian-target (setup.cfg)
 scipy==1.7.3
     # via scikit-learn
+six==1.16.0
+    # via python-dateutil
 threadpoolctl==3.0.0
     # via scikit-learn
