+++ setup.cfg
@@ -1,13 +1,12 @@
 [metadata]
-name = bayesian-target
-version = attr: bayesian_target.__version__
+name = bayte
+version = attr: bayte._meta.__version__
 description = Bayesian target encoding with scikit-learn and scipy
-license = MIT license
 long_description = file: README.md
 long_description_content_type = text/markdown
 author = Akshay Gupta
 author_email = akgcodes@gmail.com
-url = https://github.com/ak-gupta/bayesian-target
+url = https://github.com/ak-gupta/bayte
 python_requires = 
     >=3.7
 license = MIT license
@@ -19,12 +18,14 @@ classifiers =
     Programming Language :: Python :: 3 :: Only
     Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
 
 [options]
 zip_safe = False
 include_package_data = True
 packages = find:
 install_requires = 
+    pandas>=1.0.0,<=1.3.4
     scikit-learn>=1.0.0,<=1.0.1
 
 [options.extras_require]
@@ -47,6 +48,7 @@ tests =
     pytest
     pytest-cov
 dev =
+    %(build)s
     %(docs)s
     %(qa)s
     %(tests)s
@@ -58,7 +60,7 @@ version_pattern = "MAJOR.MINOR.PATCH"
 [bumpver:file_patterns]
 setup.cfg = 
     current_version = "{version}"
-bayesian_target/__init__.py = 
+bayte/_meta.py = 
     __version__ = "{version}"
 
 [flake8]
